Ai_Issue_Tracker/requirements.txt
# === Core ===
requests>=2.31.0
python-dotenv>=1.0.1
pyyaml>=6.0.2

# === Crawlers ===
beautifulsoup4>=4.12.3
lxml>=5.1.0
feedparser>=6.0.11
telethon>=1.36.0         # Telegram MTProto client
playwright>=1.40.0       # X 웹 크롤링 + JS 렌더링 사이트
requests-oauthlib>=1.3.1 # (optional, X OAuth)

# === Summarizer ===
openai>=1.40.0        # OpenAI API (Codex fallback)

# === Bot ===
python-telegram-bot>=21.0

# === Scheduler ===
APScheduler>=3.10.4

# === Optional: JS-rendered sites ===
# playwright>=1.40.0  # uncomment if needed

# === Optional: timezone backport for Python < 3.9 ===
# backports.zoneinfo>=0.2.1
